Add Switchy UI integration.
