+++ gradle/libs.versions.toml
@@ -4,7 +4,7 @@ minecraft = "1.19.2"
 quilt_mappings = "1.19.2+build.19"
 quilt_loader = "0.17.4"
 
-switchy = "2.0.0-beta.12+1.19"
+switchy = "2.0.0-beta.13+1.19"
 
 quilted_fabric_api = "4.0.0-beta.14+0.62.0-1.19.2"
 
@@ -14,6 +14,8 @@ quilt_mappings = { module = "org.quiltmc:quilt-mappings", version.ref = "quilt_m
 quilt_loader = { module = "org.quiltmc:quilt-loader", version.ref = "quilt_loader" }
 
 switchy = { module = "folk.sisby:switchy-core", version.ref = "switchy" }
+switchy_client = { module = "folk.sisby:switchy-client", version.ref = "switchy" }
+switchy_ui = { module = "folk.sisby:switchy-ui", version.ref = "switchy" }
 
 quilted_fabric_api = { module = "org.quiltmc.quilted-fabric-api:quilted-fabric-api", version.ref = "quilted_fabric_api" }
 quilted_fabric_api_deprecated = { module = "org.quiltmc.quilted-fabric-api:quilted-fabric-api-deprecated", version.ref = "quilted_fabric_api" }
