+++ src/main/java/xyz/fulmine/switchy_teleport/modules/LocationModule.java
@@ -1,30 +1,17 @@
 package xyz.fulmine.switchy_teleport.modules;
 
 import folk.sisby.switchy.api.SwitchyEvents;
-import folk.sisby.switchy.api.module.SwitchyModule;
-import folk.sisby.switchy.api.module.SwitchyModuleEditable;
-import folk.sisby.switchy.api.module.SwitchyModuleInfo;
-import folk.sisby.switchy.api.module.SwitchyModuleRegistry;
-import net.minecraft.nbt.NbtCompound;
-import net.minecraft.nbt.NbtElement;
+import folk.sisby.switchy.api.module.*;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
-import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.Nullable;
 import xyz.fulmine.switchy_teleport.Location;
-import xyz.fulmine.switchy_teleport.SwitchyTeleport;
 import xyz.fulmine.switchy_teleport.utils.TeleportUtils;
 
-public class LocationModule implements SwitchyModule, SwitchyEvents.Init {
-	final static Identifier ID = new Identifier(SwitchyTeleport.ID, "last_location");
-
-	final static String KEY_LOCATION = "last_location";
-
-	@Nullable private Location location = null;
-
+public class LocationModule extends LocationModuleData implements SwitchyModule, SwitchyModuleTransferable, SwitchyEvents.Init {
 	@Override
 	public void updateFromPlayer(ServerPlayerEntity player, @Nullable String nextPreset) {
-		this.location = new Location(
+		location = new Location(
 				player.getX(), player.getY(), player.getZ(),
 				player.getPitch(), player.getYaw(),
 				player.getWorld().getRegistryKey().getValue()
@@ -34,46 +21,8 @@ public class LocationModule implements SwitchyModule, SwitchyEvents.Init {
 	@Override
 	public void applyToPlayer(ServerPlayerEntity player) {
 
-		if (this.location != null) {
-			TeleportUtils.teleportPlayer(player, this.location);
-		}
-	}
-
-	@Override
-	public NbtCompound toNbt() {
-		NbtCompound outNbt = new NbtCompound();
-
-		NbtCompound nbtLast = new NbtCompound();
-
-		if (this.location != null) {
-			nbtLast.putDouble("x", this.location.getX());
-			nbtLast.putDouble("y", this.location.getY());
-			nbtLast.putDouble("z", this.location.getZ());
-
-			nbtLast.putFloat("pitch", this.location.getPitch());
-			nbtLast.putFloat("yaw", this.location.getYaw());
-
-			nbtLast.putString("dimension", this.location.getDimensionID().toString());
-		}
-
-		outNbt.put(KEY_LOCATION, nbtLast);
-
-		return outNbt;
-	}
-
-	@Override
-	public void fillFromNbt(NbtCompound nbt) {
-		if (nbt.contains(KEY_LOCATION, NbtElement.COMPOUND_TYPE)) {
-			NbtCompound nbtLast = nbt.getCompound(KEY_LOCATION);
-			if (nbtLast.contains("x", NbtElement.DOUBLE_TYPE)) {
-				this.location = new Location(
-						nbtLast.getDouble("x"),
-						nbtLast.getDouble("y"), nbtLast.getDouble("z"),
-						nbtLast.getFloat("pitch"),
-						nbtLast.getFloat("yaw"),
-						new Identifier(nbtLast.getString("dimension"))
-				);
-			}
+		if (location != null) {
+			TeleportUtils.teleportPlayer(player, location);
 		}
 	}
