+++ src/main/resources/assets/switchy_teleport/lang/en_us.json
@@ -1,0 +1,15 @@
+{
+  "lib39:enable_enhanced_lang": true,
+  "switchy": {
+    "teleport": {
+      "module": {
+        "last_location": {
+          "tooltip": "Currently at X:{} Y:{} Z:{}"
+        },
+        "spawn_point": {
+          "tooltip": "Spawns at X:{} Y:{} Z:{}"
+        }
+      }
+    }
+  }
+}
